+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added new heatmap chart type for GitHub-style contribution/activity charts
+
 ## [0.3.0] - 2025-02-25
 
 ### Changed